// apps/website-example/src/app/types.ts
export interface ImageData {
  url: string
  width: number
  height: number
  description?: string
  caption?: string
  source?: string
  focusPoint?: {x: number; y: number}
  format: string

  ogURL: string
  smallTeaserURL: string
  mediumTeaserURL: string
  largeURL: string
}